Add 72 hour sla to existing tests
