+++ cla_backend/apps/call_centre/tests/test_utils.py
@@ -1,7 +1,10 @@
+import datetime
 from unittest import TestCase
-
+import mock
 import jsonpatch
 from core.utils import format_patch
+from cla_common.call_centre_availability import OpeningHours
+from legalaid.utils import sla
 
 
 class FormatPatchTestCase(TestCase):
@@ -19,3 +22,20 @@ class FormatPatchTestCase(TestCase):
         formatted = format_patch(patch)
         self.assertIsInstance(formatted, basestring)
         self.assertEqual(formatted, "Changed foo to 2")
+
+
+class CallCentreFixedOperatingHours(object):
+    def setUp(self):
+        super(CallCentreFixedOperatingHours, self).setUp()
+
+        hours = {
+            "weekday": (datetime.time(9, 0), datetime.time(20, 0)),
+            "saturday": (datetime.time(9, 0), datetime.time(12, 30)),
+        }
+        operator_hours = OpeningHours(**hours)
+
+        self.operator_hours_patcher = mock.patch.object(sla, "operator_hours", operator_hours)
+        self.operator_hours_patcher.start()
+
+    def tearDown(self):
+        self.operator_hours_patcher.stop()
